Support exception handling to access tracing context
