+++ agent/agent-plugins/spring-scheduling/src/main/java/org/bithon/agent/plugin/spring/scheduling/ScheduledMethodRunnable$Run.java
@@ -19,18 +19,10 @@ package org.bithon.agent.plugin.spring.scheduling;
 import org.bithon.agent.bootstrap.aop.AbstractInterceptor;
 import org.bithon.agent.bootstrap.aop.AopContext;
 import org.bithon.agent.bootstrap.aop.InterceptionDecision;
-import org.bithon.agent.core.context.AgentContext;
-import org.bithon.agent.core.tracing.Tracer;
-import org.bithon.agent.core.tracing.config.TraceConfig;
 import org.bithon.agent.core.tracing.context.ITraceContext;
-import org.bithon.agent.core.tracing.context.TraceContextFactory;
+import org.bithon.agent.core.tracing.context.ITraceSpan;
 import org.bithon.agent.core.tracing.context.TraceContextHolder;
-import org.bithon.agent.core.tracing.propagation.TraceMode;
-import org.bithon.agent.core.tracing.sampler.ISampler;
-import org.bithon.agent.core.tracing.sampler.SamplerFactory;
-import org.bithon.agent.core.tracing.sampler.SamplingMode;
 import org.bithon.component.commons.tracing.SpanKind;
-import org.springframework.scheduling.support.ScheduledMethodRunnable;
 
 /**
  * {@link org.springframework.scheduling.support.ScheduledMethodRunnable#run()}
@@ -40,59 +32,34 @@ import org.springframework.scheduling.support.ScheduledMethodRunnable;
  */
 public class ScheduledMethodRunnable$Run extends AbstractInterceptor {
 
-    private ISampler sampler;
-
-    @Override
-    public boolean initialize() {
-        TraceConfig traceConfig = AgentContext.getInstance()
-                                              .getAgentConfiguration()
-                                              .getConfig(TraceConfig.class);
-        TraceConfig.SamplingConfig samplingConfig = traceConfig.getSamplingConfigs().get("spring-scheduler");
-        if (samplingConfig == null || samplingConfig.isDisabled() || samplingConfig.getSamplingRate() == 0) {
-            return false;
-        }
-
-        sampler = SamplerFactory.createSampler(samplingConfig);
-
-        return true;
-    }
-
+    /**
+     * The {@link org.springframework.scheduling.support.ScheduledMethodRunnable} is delegated to {@link org.springframework.scheduling.support.DelegatingErrorHandlingRunnable}.
+     * And the interceptor for the latter one is responsible for setting up the tracing context.
+     */
     @Override
     public InterceptionDecision onMethodEnter(AopContext aopContext) {
-        ITraceContext context;
-        SamplingMode mode = sampler.decideSamplingMode(null);
-        if (mode == SamplingMode.NONE) {
+        ITraceContext context = TraceContextHolder.current();
+        if (context == null) {
             return InterceptionDecision.SKIP_LEAVE;
-        } else {
-            // create a traceable context
-            context = TraceContextFactory.create(TraceMode.TRACE,
-                                                 Tracer.get().traceIdGenerator().newTraceId());
         }
 
-        ScheduledMethodRunnable runnable = aopContext.castTargetAs();
-        context.currentSpan()
-               .component("springScheduler")
-               .kind(SpanKind.TIMER)
-               .method(aopContext.getMethod())
-               .tag("uri", "spring-scheduler://" + runnable.getTarget().getClass().getName())
-               .start();
-
-        aopContext.setUserContext(context);
-        TraceContextHolder.set(context);
+        aopContext.setUserContext(context.currentSpan()
+                                         .component("springScheduler")
+                                         .kind(SpanKind.TIMER)
+                                         .method(aopContext.getMethod())
+                                         .tag("uri", "spring-scheduler://" + aopContext.getTargetClass().getName())
+                                         .start());
 
         return InterceptionDecision.CONTINUE;
     }
 
     @Override
     public void onMethodLeave(AopContext aopContext) {
-        ITraceContext context = aopContext.castUserContextAs();
+        ITraceSpan span = aopContext.getUserContextAs();
 
-        context.currentSpan()
-               .tag(aopContext.getException())
-               .tag("status", aopContext.hasException() ? "500" : "200")
-               .finish();
-        context.finish();
-
-        TraceContextHolder.remove();
+        span.tag(aopContext.getException())
+            .tag("status", aopContext.hasException() ? "500" : "200")
+            .finish();
+        span.context().finish();
     }
 }
